FR29338: Read SIM Contacts from EF-ADN if EF-PBR read failed.
Currently on a USIM UICC card, contacts are read from EF-PBR.
If EF-PBR file is absent or the read failed, no contacts displayed.
This requirement will add the support for reading SIM contacts from
EF-ADN, if EF-PBR read failed.

Fix: In IccPhoneBookInterfaceManager, while loading the records check
for EF-PBR present. If EF-PBR is absent, treat the card as 2g and
reload the records from EF-ADN.

Add code to check if EF-PBR is present in UsimPhoneBookManager.

Change-Id: I55ef0541789cd5958da26cd18b070a46d75d78b0
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>