Call Forwarding Icon displayed for Deactivated SUB
Call forwarding icon is displayed in the notification bar for the
deactivated SUB after reboot. Since the icc records are null
for the SIM it reads the status from the preference to display the
icon. An additional check is added to identify whether the SIM
is active or not.

CRs-Fixed: 943338

Change-Id: I4f9ec1fab40d39e1cc6a8611b38192ba248eb10f
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>