Telephony: Pass the subId when set data enabled
After rebooting device, setDataEnabled() is called using
the dummy subId, but before CMD_SET_USER_DATA_ENABLE is handled,
the vaild subId is read, so in onSetUserDataEnabled() method we use
valid subId to update db but not the dummy subId which we want to set.

Pass the subId when set data enabled to ensure that we can
save the mobile_data value to db using the correct subId.

Change-Id: I7c3944fa4c9922e66170e320742371374a034163
CRs-fixed: 930649
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>