Move main thread in IPBIM to other thread to fix deadlock.
When thread A waits to get result A from main thread in
IccPhoneBookInterfaceManager (IPBIM), its lock in IPBIM is released.
So thread B holds this lock to query PB to post message to main
thread in UsimPhoneBookManager to get result B. Then result A
returns into main thread in IPBIM, it acquires lock in IPBIM to
notify thread A, but this lock in IPBIM is held by thread B.
Thread B also waits to get result B from main thread in
UsimPhoneBookManager, deadlock occurs.

Change-Id: Id6c6b6231618378c27216dbf0e13a665cbae1022
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>